Rename detail folder in redux into airlines folder
